Modificaciónn de la clase usuario
